+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New option `--ignore` to ignore errors in file and continue loading (ER #46)
 - New option `--truncate` to empty (truncate) the target table before the load (ER #45)
 - New option `--encoding` to specify the file encoding, if not compatible with `UTF-8` (bug #50)
+- New option `--case-insensitive-identifiers` to uppercase all identifiers (ER #60)
 
 ### Changed
 - Read all input files with `UTF-8` encoding by default